Issue #2228 "Grizzly NPE in DefaultFilterChain, this.serverRequest is null" (https://github.com/eclipse-ee4j/glassfish-grizzly/issues/2228) Issue #2214 "NPE when closing OutputBuffer" (https://github.com/eclipse-ee4j/glassfish-grizzly/issues/2214)
+ Fixed an issue where a new request was unintentionally processed as a suspend request because it was referencing a recycled request.
+ Improvements to the use of recycled InputBuffer and OutputBuffer and concurrent access to them.
When a request and response are completed and recycled, the associated ServerInputBuffer and ServerOutputBuffer are also recycled.
However, if another thread references the InputBuffer and OutputBuffer, an NPE may occur when using these recycled instances.
We consider the possibility of unintended usage after recycling, and the possibility of concurrent calls to recycle and use.
I did not modify InputBuffer and OutputBuffer to be completely thread-safe; Improvements have been made only to the areas mentioned in the issues above.
+ Patched an NPE that occurred when calling toString() on a recycled HttpRequestPacket instance (preventing an NPE when obtaining the methodBytes of Method).
+ Patched for TransportFilter's exceptionOccurred() not propagating to filter chain (it seems that there was no apparent intention to prevent propagation).